+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,7 +6,7 @@
 
   <fragment
     android:id="@+id/navigation_home"
-    android:name="com.waffiq.bazz_movies.feature.home.ui.HomeFragment"
+    android:name="com.waffiq.bazz_movies.feature.home.ui.fragment.HomeFragment"
     android:label="@string/title_home"
     tools:layout="@layout/fragment_featured" />
