+++ feature/home/src/main/res/layout/fragment_movie.xml
@@ -5,7 +5,7 @@
   android:id="@+id/coordinator_layout"
   android:layout_width="match_parent"
   android:layout_height="match_parent"
-  tools:context=".ui.MovieFragment">
+  tools:context=".ui.fragment.MovieFragment">
 
   <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
     android:id="@+id/swipe_refresh_movie"
